Add files via upload
Attendance System Model: Description: In this task you must make machine learning models to detect and identify students in a class. If a student is detected mark present and if absent mark absent. It should detect their emotions too. At the end final data should be stored in a excel or csv file with time of identification. This model should only work for a limited time like 9.30 AM to 10.00 AM. Guidelines: You should train your own machine learning model. GUI for this task is not mandatory. Only the model should work properly and should fulfil the above requirements.